# front/src/lib/generated/models/MonthlyStatisticsDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type MonthlyStatisticsDto = {
    productName?: string;
    totalBought?: number;
    totalSpent?: number;
    totalSold?: number;
    totalProfit?: number;
    stockAtEndOfMonth?: number;
};