# front/src/lib/generated/models/Product.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type Product = {
    id?: number;
    name?: string;
    currentStock?: number;
    lastBoughtPrice?: number;
    lastSoldPrice?: number;
};