# front/src/lib/generated/models/Sellings.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { Product } from './Product';
export type Sellings = {
    id?: number;
    product?: Product;
    quantity?: number;
    price?: number;
    totalPrice?: number;
    sellingDate?: string;
};