# front/src/lib/generated/models/TransactionDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type TransactionDto = {
    productId?: number;
    quantity?: number;
    price?: number;
    transactionDate?: string;
};