# ms-coffeeShop/ms-coffeeShop/src/main/java/com/ms_coffeeShop/entity/Purchases.java
package com.ms_coffeeShop.entity;

import jakarta.persistence.*;
import lombok.*;
import java.time.LocalDate;

@Entity
@Data
@Table(name = "purchases")
public class Purchases {
    @Id
    @GeneratedValue(strategy = GenerationType.IDENTITY)
    private Long id;

    @ManyToOne
    @JoinColumn(name = "product_id")
    private Product product;

    private Integer quantity;
    private Double price;
    private Double totalPrice;
    private LocalDate buyingDate;
}